requirements.txt
python-binance>=1.0.19
yfinance>=0.2.37
pandas>=2.1.4
